Add support for unittests
This brings in googletest as an external dependency, builds it in a
non-sandboxed environment and produces a library that can be linked to
as well as exposing headers that can be used by the bazel project.

We've had to make a minor change to one cpp file, but this was preferable
to manually putting the tap.h file in googletest.